refactor: centralize password utilities
