+++ auth-service/src/service/auth.service.ts
@@ -1,6 +1,6 @@
 import { prisma } from '../core/prisma'
 import jwt from 'jsonwebtoken'
-import { hashPassword, comparePassword } from '../util/password'
+import { hashPassword, comparePassword } from '../../../shared/core/password'
 
 const JWT_SECRET = process.env.JWT_SECRET!
 const JWT_EXPIRES_IN = '7d'
